Use append_cflags instead of modifying CFLAGS directly
This fixes a compilation issue in GCC 15 due `stdbool.h`
not being included when `-std=c99` is in use. This occurs because
the standard `configure` script shipped with Ruby does not properly
test for `stdbool.h` on C23 compilers. It needs to pick up this change:

https://cgit.git.savannah.gnu.org/cgit/autoconf.git/commit/?id=6dcecb780a69bd208088d666b299e92aa7ae7e80

The `append_cflags` approach is recommended over modifying CFLAGS since
it ensures compatibility across different build environments by checking
whether the flag is acceptable.

Similar changes:

* https://github.com/socketry/io-event/pull/137
* https://github.com/mongodb/bson-ruby/pull/355

Closes #27